Rename OptionsView as CdnConfigView
