+++ admin/src/CdnConfigView.module.css
